add architecture diagram
this adds a rough diagram showing the (software) architecture.

we'll most likely use a bunch of pre-existing ROS nodes like
[diff_drive_controller][] or [rplidar][] which i haven't added yet to
the diagram (for a list of nodes which _might_ be useful to us check
out what the SMB example from ETH is using as dependencies in their
`smb_control/package.xml`).

the state diagram will also depend on the outcome of the discussion in
PR #1.

this diagram is written using the [mermaid][] syntax which can be
rendered by [GitHub][gh-mermaid] as well as various IDEs through plugins
(e.g. the [JetBrains Mermaid plugin][jb-mermaid] which i'm using in
CLion).

[diff_drive_controller]: https://wiki.ros.org/diff_drive_controller
[rplidar]: https://wiki.ros.org/rplidar
[mermaid]: https://mermaid.js.org/
[gh-mermaid]: https://docs.github.com/en/repositories/working-with-files/using-files/working-with-non-code-files#displaying-mermaid-files-on-github
[jb-mermaid]: https://plugins.jetbrains.com/plugin/20146-mermaid