enable GitHub dependabot for GitHub Actions
this enables the version updates for the GitHub Actions. note that we
can't do the same for the C++ code yet because (a) we don't have any
yet, (b) dependabot doesn't support ROS dependencies and (c) ROS
dependencies anyway don't seem to specify a version.

see [the docs][] for further details.

[the docs]: https://docs.github.com/en/code-security/dependabot/dependabot-version-updates