Add a new model for the smt in gazebo
This commit adds a new Gazebo model for the SMT robot that allows for testing of its features in simulation without requiring a physical robot. To reduce the computational load of the simulation, the model replaces the two tracks of the robot with two wheels on each side. The exact torque and traction of the wheels will need to be measured on the physical robot and adjusted in the model if necessary.