add `optional` library as git submodule
C++17 would come with [`std::optional`], but due to ROS 1 we're stuck on
C++14. to make the code cleaner we nevertheless want to rely on an
`optional` in the API. Thankfully, we can use the header-only library
[`tl::optional`] as a replacement.

As we don't have vcpkg easily available in this environment we instead
use a [git submodule] to include it - we can then just set the correct
include path in the other project(s) to make use of it.

important: as this adds a new submodule (the first in this project)
everyone working on it must init the submodule:
```bash
git submodule init
git submodule update
```

[`std::optional`]: https://en.cppreference.com/w/cpp/utility/optional
[`tl::optional`]: https://github.com/TartanLlama/optional
[git submodule]: https://git-scm.com/book/en/v2/Git-Tools-Submodules