add path finder algorithm - WIP
this doesn't work correctly yet and needs further refinement. the
problem is that if we're too close we just try to head to the farthest
point - but that can change at any time.
we should instead remember our previous heading and then try to head in
that direction until we get too close to things.

TODOs:
* fix algorithm
* rename variables to camelCase
* add documentation
* add test coverage
* remove `extractValidMeasurements`