add dependency to `pigpio` library in CMake build
use CMake's [`ExternalProject`] functionality to pull in the [`pigpio`]
library.

we explicitly check out the latest release tag (instead of just the
default branch) to ensure that we always have the same version.
the installation directory is explicitly specified as otherwise it tries
to install the library globally, which would require root access. as we
only need it for our build we don't want to interfer with the rest of
the system and don't need root.

in the best case this would've been mostly provided by the library
directly, but so far this isn't the case. two issues,
joan2937/pigpio#572 and joan2937/pigpio#573 have been raised by me while
implementing this to hopefully make this better in the future.

[`ExternalProject`]: https://cmake.org/cmake/help/latest/module/ExternalProject.html
[`pigpio`]: https://github.com/joan2937/pigpio