add `better_teleop` as git submodule
this adds the [`better_teleop`] package which provides keyboard
operation for the robot. it currently has to be added as a git submodule
as it is not (yet) published in such a way that it could be installed
using `rosdep install` (this might eventually come, but probably too
late for us; i've raised ethz-asl/better_teleop#3 for this).

in order to use this properly as a quick-win we need to manually run
`sudo pip install pynput` to install the required dependencies.
this should hopefully improve slightly in the future, either due to
ethz-asl/better_teleop#1 or ethz-asl/better_teleop#2.

important: as this adds a new submodule (the first in this project)
everyone working on it must init the submodule:
```bash
git submodule init
git submodule update
```

[git submodule]: https://git-scm.com/book/en/v2/Git-Tools-Submodules
[`better_teleop`]: https://github.com/ethz-asl/better_teleop