use `better_teleop` in `smt_launch_gazebo`
with this it's now possible to use the keyboard to navigate the robot in
the world.
we'll also keep this when we add a controller with an algorithm to drive
around as we'll be able to turn the latter on/off (using a ROS service
call) and can then directly take over with the keyboard (when the
service is on it'll issue a continous stream of messages on `/cmd_vel`,
overwriting anything `better_teleop` would issue).