add `smt_launch_hardware`
with this running `roslaunch smt_launch_hardware default.launch` will
launch all required nodes to run the logic on the actual hardware (a
Raspberry Pi 4).

note that this does not launch rviz, teleop or anything else as it makes
no sense to run this on the robot. instead, you can set your
`ROS_MASTER_URI` on your own device to point at the robot (see the
output of `roscore` when starting it) and then start these nodes on your
device to interact with the robot if needed.

important: since i don't have the hardware available i'm unable to test
this myself!