add `rplidar_ros` for HW launch
the [`rplidar_ros`] package is used to read the data from the physical
LIDAR and create ROS messages out of it. this package is provided by the
manufacturer of the LIDAR.

note that after this commit you need to run
```bash
rosdep install --from-paths . -i
```
again. and on RPi used on the robot you once need to run the script to
set up the new udev rules (after `rosdep install` as it needs
`rplidar_ros` to be present):
```bash
$(rospack find rplidar_ros)/scripts/create_udev_rules.sh
```

[`rplidar_ros`]: https://wiki.ros.org/rplidar