add initial SLAM functionality
this makes use of [`hector_slam`] for the actual SLAM implementation.

the quality of the map in the initial test wasn't that great, but this
might be because the robot moved quite fast and various articles only
said that for hector SLAM to work properly it needs to turn slowly.

important: this hasn't been tested yet on the RPi itself. it might be
that some additional configuration is needed there to publish the
necessary robot state transformations (in the simulation we have this in
the launch file already and gazebo also takes care of some of it).
also, this will require #49 (which adds rplidar) for it to work on the
robot as otherwise we won't have any laser scan messages there.

[`hector_slam`]: https://wiki.ros.org/hector_slam