add ekf localization for odom->base_link transform
this adds a `ekf_localization_node` (from the [`robot_localization`]
package) so that we can have an odometry frame.

since we don't have a wheeled robot it's hard to have wheel odometry
(though we could probably do something with our tracks as well, but this
is less common).

instead, we use the data from our newly added IMU and also fuse it with
the pose estimate created by our mapping tool (see `poseupdate` in the
[`hector_mapping`] documentation).

at the moment this seems to have quite a bit of drift when creating a
map, some fine-tuning might be possible.

[`robot_localization`]: https://wiki.ros.org/robot_localization
[`hector_mapping`]: https://wiki.ros.org/hector_mapping#Published_Topics