HW launch: use `lidar_1` as `frame_id`
by default the `rplidar.launch` sets the `frame_id` to `laser`, which is
different from what we've set in our model and what we use in our
configuration. accordingly, the mapping algorithm so far failed to
properly handle the laser scan messages.

as `rplidar.launch` doesn't expose this as an argument i instead just
copied over the whole node launch code and adjusted this one line.