rewrite `pigpio` code to `pigpiod_if2`
we currently start a `pigpiod` as part of our own process. this has
several drawbacks:
* "there can be only one!": we cannot have two packages interacting with
  the GPIO pins separately (e.g. one the movement & gun controllers)
* there's already a `pigpiod` running on the default port 8888, so we
  currently have to manually set ours to 8889
* if our application doesn't terminate properly we have to kill it
  manually to be able to start it again
* we need to launch our node which uses `pigpio` with `sudo` for it to
  work

instead of using our current approach with [the C interface][pigpio-cif]
we can use [pigpiod_if2]. this is an API which connects to an already
running deamon (and in our case it's always already running as it's
started as a service) and sends the commands there.
this allows us to start an arbitrary amount of nodes which talk to the
GPIO pins and they don't need `sudo` rights.

[pigpio-cif]: https://abyz.me.uk/rpi/pigpio/cif.html
[pigpiod_if2]: http://abyz.me.uk/rpi/pigpio/pdif2.html