README: document features
note: the documentation for the QR code scanner covers features which
have not yet been implemented / integrated. if this is not available
when we have the presentation we should update the README again and
remove it.

in turn, if we add e.g. a self-driving mode or a "drive towards" mode
the README needs to be updated to reflect this.