implement gun controller
this adds a basic gun controller which will run on the physical hardware
and allow moving the gun up/down and firing it.

this has been ported from the original PR from Joel to the new approach
based on `pigpiod_if2` and some further cleanups applied.

this currently uses a ROS topic, however this would probably be more
suited to a ROS service based on how this is supposed to be used.
for time reasons i haven't rewritten it to that at this point.

Co-Authored-By: Joel Flepp <joel_flepp@hotmail.com>