gazebo: set `gaussianNoise` for IMU
the simulated IMU sends relatively bad data (e.g. the z acceleration
isn't fixed at 9.81 but fluctuates wildely around that number). however,
so far the covariance in the IMU data was always set to 0.0 which meant
that the `ekf_localization_node` thought that the data was perfect and
used it as the primary source, thus leading to wrong position estimates
for the `odom` frame.
by setting the `gaussianNoise` it now sets a covariance and the
localization algorithm gives the IMU data slightly less weight.

the number 0.5 was arrived at experimentally, it seems to give a good
result: this allows mapping the whole map in gazebo without any
noticeable drift.