add framework for QR code detection
this adds the necessary ROS infrastructure to detect QR codes, calculate
their position and publish this information so that it can be consumed
by another node (which will be added afterwards and which will publish
commands to the gun to fire it at the target, possibly also at the robot
to move it into place).

the `smt_qr_msgs` package contains the message structure(s) used to
publish the QR code location and related information (namely the content
of the QR code).
the location is given as a pose so that the robot can also orient itself
to properly hit the QR code if needed.

the `smt_qr_finder` scans the images for QR codes and publishes a
message for each QR code being found and also publishes a copy of the
image with the QR code tagged in case at least one QR code has been
found.
note that at the moment the algorithm to calculate the position of the
QR code has not yet been implemented!

to find the QR codes the [zbar library] is used. OpenCV would also have
provided this functionality, however it was not present in the compiled
version being used by ROS and we didn't find an easy way to force it to
use another OpenCV version which would include it.

Co-Authored-By: Joel Flepp <joel_flepp@hotmail.com>

[zbar library]: https://zbar.sourceforge.net/