suppress annoying 3rd party CMake warning
the warning occurs because these projects still use a CMake version 2.x
specifier in their `CMakeLists.txt` which had a different policy than
CMake 3.x which we're using. however, the warning being shown is only
useful to the maintainers of these libraries and not to us, so we can
follow the recommendation (printed by the warning!) to suppress it in
this way.